Moved group type classes to separate namespace
